app2/requirements .txt
Flask==2.0.1
SQLAlchemy==1.4.23
Pillow==8.3.1
Flask-Login==0.5.0
requests==2.25.1
